+++ src/components/atoms/action-card/ActionCard.stories.ts
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import { ActionCard } from '@/components/molecules/action-card/ActionCard';
+import { ActionCard } from '@/components/atoms/action-card/ActionCard';
 
 // More on how to set up stories at: https://storybook.js.org/docs/writing-stories#default-export
 const meta = {
