+++ src/components/atoms/hero/Hero.stories.ts
@@ -1,17 +1,17 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import { LoadingProjectCard } from '@/components/molecules/loading-project-card/LoadingProjectCard';
+import { Hero } from '@/components/atoms/hero/Hero';
 
 // More on how to set up stories at: https://storybook.js.org/docs/writing-stories#default-export
 const meta = {
-  title: 'OpendataBridge/Molecules/LoadingProjectCard',
-  component: LoadingProjectCard,
+  title: 'OpendataBridge/Molecules/Hero',
+  component: Hero,
   parameters: {
     // Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/configure/story-layout
-    layout: 'center',
+    layout: 'fullscreen',
   },
   // This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/writing-docs/autodocs
   tags: ['autodocs'],
-} satisfies Meta<typeof LoadingProjectCard>;
+} satisfies Meta<typeof Hero>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
