+++ src/components/atoms/resource-option/ResourceOption.stories.ts
@@ -1,4 +1,4 @@
-import { ResourceOption } from '@/components/molecules/resource-option/ResourceOption';
+import { ResourceOption } from '@/components/atoms/resource-option/ResourceOption';
 import { Meta, type StoryObj } from '@storybook/react';
 
 const meta = {
