+++ src/components/molecules/modal/success-formating-modal/success-formatting-modal.scss
