+++ src/components/molecules/project/dataset-card/DatasetCard.stories.ts
@@ -1,24 +1,27 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import { ProjectCard } from '@/components/molecules/project-card/ProjectCard';
-import { ProjectTags } from '@/components/molecules/project-tags/ProjectTags';
+import { DatasetCard } from '@/components/molecules/project/dataset-card/DatasetCard';
 
 // More on how to set up stories at: https://storybook.js.org/docs/writing-stories#default-export
 const meta = {
-  title: 'OpendataBridge/Molecules/ProjectTags',
-  component: ProjectTags,
+  title: 'OpendataBridge/Molecules/DatasetCard',
+  component: DatasetCard,
   parameters: {
     // Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/configure/story-layout
     layout: 'center',
   },
   // This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/writing-docs/autodocs
   tags: ['autodocs'],
-} satisfies Meta<typeof ProjectTags>;
+} satisfies Meta<typeof DatasetCard>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
 
 export const Default: Story = {
   args: {
-    tags: ['tag1', 'tag2', 'tag3'],
+    dataset: {
+      title: 'ファイル名',
+      organization: '自治体名',
+      url: '',
+    },
   },
 };
