+++ src/components/templates/detail-project/DetailProject.tsx
@@ -1,10 +1,10 @@
 'use client';
 import React from 'react';
 import { Project } from '@/types/project';
-import { ProjectCard } from '@/components/molecules/project-card/ProjectCard';
-import { DatasetCard } from '@/components/molecules/dataset-card/DatasetCard';
-import { ActionCard } from '@/components/molecules/action-card/ActionCard';
-import { ProjectTags } from '@/components/molecules/project-tags/ProjectTags';
+import { ProjectCard } from '@/components/molecules/project/project-card/ProjectCard';
+import { DatasetCard } from '@/components/molecules/project/dataset-card/DatasetCard';
+import { ActionCard } from '@/components/atoms/action-card/ActionCard';
+import { ProjectTags } from '@/components/molecules/project/project-tags/ProjectTags';
 import CodeEditor from '@/components/atoms/code-editor/CodeEditor';
 
 interface DetailProjectProps {
