+++ src/components/templates/develop-recipe/DevelopRecipe.tsx
@@ -1,13 +1,13 @@
 'use client';
 import React from 'react';
 import { Project } from '@/types/project';
-import { ProjectCard } from '@/components/molecules/project-card/ProjectCard';
-import { ProjectTags } from '@/components/molecules/project-tags/ProjectTags';
+import { ProjectCard } from '@/components/molecules/project/project-card/ProjectCard';
+import { ProjectTags } from '@/components/molecules/project/project-tags/ProjectTags';
 import { Button } from '@/components/atoms/button/Button';
 import { useFileList } from '@/hooks/use-file-list';
 import { InputFileList } from '@/components/organizms/input-file-list/InputFileList';
 import { Alert } from '@/components/atoms/alert/Alert';
-import { TableView } from '@/components/molecules/table/TableView';
+import { TableView } from '@/components/atoms/table/TableView';
 import InputRecipe from '@/components/molecules/input-recipe/InputRecipe';
 import { useInputPrompt } from '@/hooks/use-input-prompt';
 import CodeEditor from '@/components/atoms/code-editor/CodeEditor';
