+++ src/hooks/use-input-prompt.ts
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import useSWR from 'swr';
-import { GenericDataType } from '@/components/molecules/table/TableView';
+import { GenericDataType } from '@/components/atoms/table/TableView';
 import useSWRMutation from 'swr/mutation';
 import { Recipe } from '@/types/recipe';
